refactor: Move and reconfigure workbooks to features (#3193)
