+++ src/routes/workbooks/+page.svelte
@@ -15,16 +15,16 @@
     type WorkbooksList,
     type WorkBookTaskBase,
     WorkBookType,
-  } from '$lib/types/workbook';
+  } from '$features/workbooks/types/workbook';
 
-  import { activeWorkbookTabStore } from '$lib/stores/active_workbook_tab';
+  import { activeWorkbookTabStore } from '$features/workbooks/stores/active_workbook_tab';
 
   import HeadingOne from '$lib/components/HeadingOne.svelte';
   import TabItemWrapper from '$lib/components/TabItemWrapper.svelte';
-  import WorkBookList from '$lib/components/WorkBooks/WorkBookList.svelte';
+  import WorkBookList from '$features/workbooks/components/list/WorkBookList.svelte';
 
   import { calcGradeMode } from '$lib/utils/task';
-  import { canViewWorkBook } from '$lib/utils/workbooks';
+  import { canViewWorkBook } from '$features/workbooks/utils/workbooks';
 
   let { data } = $props();
