+++ src/test/lib/zod/auth_schema.test.ts
@@ -1,5 +1,5 @@
 import { expect, test } from 'vitest';
-import { ZodSchema } from 'zod';
+import { type ZodSchema } from 'zod';
 
 import { authSchema } from '$lib/zod/schema';
